telemedicine/requirements.txt
django>=5.0.7
djangorestframework>=3.14.0
firebase-admin>=7.4.0
twilio>=9.0.0
python-dotenv>=1.0.0
